% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmcfit.R
\name{vlmcfit}
\alias{vlmcfit}
\title{Context Algorithm for fitting VLMC trees}
\usage{
vlmcfit(z, Hmax = 8, cutoff = NULL)
}
\arguments{
\item{z}{Either an integer vector with the observed sequence
or a list of integer vectors for multiple independent sequences.}

\item{Hmax}{Maximum range of the Markov Chain considered. 
Defaults to 8.}

\item{cutoff}{Cutoff value used in the Likelihood Ratio Test for
prunning.}
}
\description{
Context Algorithm for fitting VLMC trees
}
\author{
Victor Freguglia
}
